// packages/mrbd/src/mrbd/ui/private/ButtonLayout.types.ts
/**
 * Copyright (c) Meta Platforms, Inc. and affiliates.
 *
 * This source code is licensed under the Apache License, Version 2.0 found in the
 * LICENSE file in the root directory of this source tree.
 *
 */

import type { State } from '@wearables-ui-toolkit/foundation/base/Interactions';

export enum LeadingAccessoryRenderMode {
  NONE = 'none',
  ICON = 'icon',
  AVATAR = 'avatar',
}

export interface ButtonAnimationValues {
  width: number;
  scale: number;
  iconLeadingMargin: number;
  textOpacity: number;
}

export interface ButtonTargetWidthInput {
  customWidth?: number | string;
  renderMode: LeadingAccessoryRenderMode;
  hasText: boolean;
  state: State;
  alwaysShowText: boolean;
  measuredContentWidth: number;
  enforceMaxWidth: boolean;
}